requirements.txt
Pillow
customtkinter
rich
pandas
requests
selenium
webdriver-manager
